Convert sets used in entry_points to lists
Some projects (like msm) use sets in the entry_points values, like:

    entry_points={
        'console_scripts': {
            'msm=msm.__main__:main'
        }
    }

This break the json.dumps call a few lines below in metaextract.py since
sets are not serializable in json.
This commit casts those sets to lists so the data can be written to a json
file and py2pack doesn't break when generating a spec file for those projects.